Add files via upload
In this app there are 4 screens.  Screen 1 is the login screen the user of this app  will login with their credentials .In the screen 1  there is a scan button that allows to scan the QR code. Screen 2 is the signup screen the new user will creates account by signing up. Screen 3 is the Numbers Screen , it contains Emergency numbers.  Screen 4 is the main screen of the app  in which the logged  user will generates a QR code  by giving the data of the user family member and download the generated QR into mobile storage.